requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
anyio==3.6.2
    # via starlette
fastapi==0.85.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
httplib2==0.20.4
    # via oauth2
hypercorn==0.14.3
    # via -r requirements.in
hyperframe==6.0.1
    # via h2
idna==3.4
    # via anyio
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
oauth2==1.9.0.post1
    # via -r requirements.in
priority==2.0.0
    # via hypercorn
pydantic==1.10.2
    # via fastapi
pyparsing==3.0.9
    # via httplib2
python-dotenv==0.21.0
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.20.4
    # via fastapi
toml==0.10.2
    # via hypercorn
typing-extensions==4.4.0
    # via pydantic
wsproto==1.2.0
    # via hypercorn
